+++ sdk/resourcemanager/paloaltonetworksngfw/armpanngfw/certificateobjectglobalrulestack_client.go
@@ -43,7 +43,7 @@ func NewCertificateObjectGlobalRulestackClient(credential azcore.TokenCredential
 // BeginCreateOrUpdate - Create a CertificateObjectGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - name - certificate name
 //   - resource - Resource create parameters.
@@ -70,7 +70,7 @@ func (client *CertificateObjectGlobalRulestackClient) BeginCreateOrUpdate(ctx co
 // CreateOrUpdate - Create a CertificateObjectGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 func (client *CertificateObjectGlobalRulestackClient) createOrUpdate(ctx context.Context, globalRulestackName string, name string, resource CertificateObjectGlobalRulestackResource, options *CertificateObjectGlobalRulestackClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "CertificateObjectGlobalRulestackClient.BeginCreateOrUpdate"
@@ -108,7 +108,7 @@ func (client *CertificateObjectGlobalRulestackClient) createOrUpdateCreateReques
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, resource); err != nil {
@@ -120,7 +120,7 @@ func (client *CertificateObjectGlobalRulestackClient) createOrUpdateCreateReques
 // BeginDelete - Delete a CertificateObjectGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - name - certificate name
 //   - options - CertificateObjectGlobalRulestackClientBeginDeleteOptions contains the optional parameters for the CertificateObjectGlobalRulestackClient.BeginDelete
@@ -146,7 +146,7 @@ func (client *CertificateObjectGlobalRulestackClient) BeginDelete(ctx context.Co
 // Delete - Delete a CertificateObjectGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 func (client *CertificateObjectGlobalRulestackClient) deleteOperation(ctx context.Context, globalRulestackName string, name string, options *CertificateObjectGlobalRulestackClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "CertificateObjectGlobalRulestackClient.BeginDelete"
@@ -184,7 +184,7 @@ func (client *CertificateObjectGlobalRulestackClient) deleteCreateRequest(ctx co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -193,7 +193,7 @@ func (client *CertificateObjectGlobalRulestackClient) deleteCreateRequest(ctx co
 // Get - Get a CertificateObjectGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - name - certificate name
 //   - options - CertificateObjectGlobalRulestackClientGetOptions contains the optional parameters for the CertificateObjectGlobalRulestackClient.Get
@@ -236,7 +236,7 @@ func (client *CertificateObjectGlobalRulestackClient) getCreateRequest(ctx conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -253,7 +253,7 @@ func (client *CertificateObjectGlobalRulestackClient) getHandleResponse(resp *ht
 
 // NewListPager - List CertificateObjectGlobalRulestackResource resources by Tenant
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - options - CertificateObjectGlobalRulestackClientListOptions contains the optional parameters for the CertificateObjectGlobalRulestackClient.NewListPager
 //     method.
@@ -292,7 +292,7 @@ func (client *CertificateObjectGlobalRulestackClient) listCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil