+++ sdk/resourcemanager/paloaltonetworksngfw/armpanngfw/fqdnlistglobalrulestack_client.go
@@ -43,7 +43,7 @@ func NewFqdnListGlobalRulestackClient(credential azcore.TokenCredential, options
 // BeginCreateOrUpdate - Create a FqdnListGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - name - fqdn list name
 //   - resource - Resource create parameters.
@@ -70,7 +70,7 @@ func (client *FqdnListGlobalRulestackClient) BeginCreateOrUpdate(ctx context.Con
 // CreateOrUpdate - Create a FqdnListGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 func (client *FqdnListGlobalRulestackClient) createOrUpdate(ctx context.Context, globalRulestackName string, name string, resource FqdnListGlobalRulestackResource, options *FqdnListGlobalRulestackClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "FqdnListGlobalRulestackClient.BeginCreateOrUpdate"
@@ -108,7 +108,7 @@ func (client *FqdnListGlobalRulestackClient) createOrUpdateCreateRequest(ctx con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, resource); err != nil {
@@ -120,7 +120,7 @@ func (client *FqdnListGlobalRulestackClient) createOrUpdateCreateRequest(ctx con
 // BeginDelete - Delete a FqdnListGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - name - fqdn list name
 //   - options - FqdnListGlobalRulestackClientBeginDeleteOptions contains the optional parameters for the FqdnListGlobalRulestackClient.BeginDelete
@@ -146,7 +146,7 @@ func (client *FqdnListGlobalRulestackClient) BeginDelete(ctx context.Context, gl
 // Delete - Delete a FqdnListGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 func (client *FqdnListGlobalRulestackClient) deleteOperation(ctx context.Context, globalRulestackName string, name string, options *FqdnListGlobalRulestackClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "FqdnListGlobalRulestackClient.BeginDelete"
@@ -184,7 +184,7 @@ func (client *FqdnListGlobalRulestackClient) deleteCreateRequest(ctx context.Con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -193,7 +193,7 @@ func (client *FqdnListGlobalRulestackClient) deleteCreateRequest(ctx context.Con
 // Get - Get a FqdnListGlobalRulestackResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - name - fqdn list name
 //   - options - FqdnListGlobalRulestackClientGetOptions contains the optional parameters for the FqdnListGlobalRulestackClient.Get
@@ -236,7 +236,7 @@ func (client *FqdnListGlobalRulestackClient) getCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -253,7 +253,7 @@ func (client *FqdnListGlobalRulestackClient) getHandleResponse(resp *http.Respon
 
 // NewListPager - List FqdnListGlobalRulestackResource resources by Tenant
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - options - FqdnListGlobalRulestackClientListOptions contains the optional parameters for the FqdnListGlobalRulestackClient.NewListPager
 //     method.
@@ -292,7 +292,7 @@ func (client *FqdnListGlobalRulestackClient) listCreateRequest(ctx context.Conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil