+++ sdk/resourcemanager/paloaltonetworksngfw/armpanngfw/postrules_client.go
@@ -43,7 +43,7 @@ func NewPostRulesClient(credential azcore.TokenCredential, options *arm.ClientOp
 // BeginCreateOrUpdate - Create a PostRulesResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - priority - Post Rule priority
 //   - resource - Resource create parameters.
@@ -70,7 +70,7 @@ func (client *PostRulesClient) BeginCreateOrUpdate(ctx context.Context, globalRu
 // CreateOrUpdate - Create a PostRulesResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 func (client *PostRulesClient) createOrUpdate(ctx context.Context, globalRulestackName string, priority string, resource PostRulesResource, options *PostRulesClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "PostRulesClient.BeginCreateOrUpdate"
@@ -108,7 +108,7 @@ func (client *PostRulesClient) createOrUpdateCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, resource); err != nil {
@@ -120,7 +120,7 @@ func (client *PostRulesClient) createOrUpdateCreateRequest(ctx context.Context, 
 // BeginDelete - Delete a PostRulesResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - priority - Post Rule priority
 //   - options - PostRulesClientBeginDeleteOptions contains the optional parameters for the PostRulesClient.BeginDelete method.
@@ -145,7 +145,7 @@ func (client *PostRulesClient) BeginDelete(ctx context.Context, globalRulestackN
 // Delete - Delete a PostRulesResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 func (client *PostRulesClient) deleteOperation(ctx context.Context, globalRulestackName string, priority string, options *PostRulesClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "PostRulesClient.BeginDelete"
@@ -183,7 +183,7 @@ func (client *PostRulesClient) deleteCreateRequest(ctx context.Context, globalRu
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -192,7 +192,7 @@ func (client *PostRulesClient) deleteCreateRequest(ctx context.Context, globalRu
 // Get - Get a PostRulesResource
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - priority - Post Rule priority
 //   - options - PostRulesClientGetOptions contains the optional parameters for the PostRulesClient.Get method.
@@ -234,7 +234,7 @@ func (client *PostRulesClient) getCreateRequest(ctx context.Context, globalRules
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -252,7 +252,7 @@ func (client *PostRulesClient) getHandleResponse(resp *http.Response) (PostRules
 // GetCounters - Get counters
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - priority - Post Rule priority
 //   - options - PostRulesClientGetCountersOptions contains the optional parameters for the PostRulesClient.GetCounters method.
@@ -294,7 +294,7 @@ func (client *PostRulesClient) getCountersCreateRequest(ctx context.Context, glo
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	if options != nil && options.FirewallName != nil {
 		reqQP.Set("firewallName", *options.FirewallName)
 	}
@@ -314,7 +314,7 @@ func (client *PostRulesClient) getCountersHandleResponse(resp *http.Response) (P
 
 // NewListPager - List PostRulesResource resources by Tenant
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - options - PostRulesClientListOptions contains the optional parameters for the PostRulesClient.NewListPager method.
 func (client *PostRulesClient) NewListPager(globalRulestackName string, options *PostRulesClientListOptions) *runtime.Pager[PostRulesClientListResponse] {
@@ -352,7 +352,7 @@ func (client *PostRulesClient) listCreateRequest(ctx context.Context, globalRule
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -370,7 +370,7 @@ func (client *PostRulesClient) listHandleResponse(resp *http.Response) (PostRule
 // RefreshCounters - Refresh counters
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - priority - Post Rule priority
 //   - options - PostRulesClientRefreshCountersOptions contains the optional parameters for the PostRulesClient.RefreshCounters
@@ -412,7 +412,7 @@ func (client *PostRulesClient) refreshCountersCreateRequest(ctx context.Context,
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	if options != nil && options.FirewallName != nil {
 		reqQP.Set("firewallName", *options.FirewallName)
 	}
@@ -424,7 +424,7 @@ func (client *PostRulesClient) refreshCountersCreateRequest(ctx context.Context,
 // ResetCounters - Reset counters
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-09-01
+// Generated from API version 2024-01-19-preview
 //   - globalRulestackName - GlobalRulestack resource name
 //   - priority - Post Rule priority
 //   - options - PostRulesClientResetCountersOptions contains the optional parameters for the PostRulesClient.ResetCounters method.
@@ -466,7 +466,7 @@ func (client *PostRulesClient) resetCountersCreateRequest(ctx context.Context, g
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-09-01")
+	reqQP.Set("api-version", "2024-01-19-preview")
 	if options != nil && options.FirewallName != nil {
 		reqQP.Set("firewallName", *options.FirewallName)
 	}